Run collector: One Pod per node when using selectors
When node selectors are defined in a run collector, one Pod is created
per node that satisfies the nodeSelector terms.

In case of needing a Pod to run in one Node specificly, the user must
define a nodeSelector that is only fulfiled by one node.

Logs from the Pod are gathered in one action, hence a timeout defined in
the run collector applies for the gathering of data from all Pods