feat: collecting and analyzing host packages
added a collector and an analyzer for packages installed in the
operating system.

here is an example of collecting installing packages:

```yaml
apiVersion: troubleshoot.sh/v1beta2
kind: HostPreflight
metadata:
  name: check-installed-packages
spec:
  collectors:
    - installedPackages:
        collectorName: "Installed packages"
```

for analyzing this pr also introduces the concept of a templated
analyzer. here is an analyzer to check a given package is present
but only on an ubuntu installation:

```yaml
spec:
  analyzers:
    - installedPackage:
        collectorName: "Installed packages"
        packageName: "dbus-user-session"
        exclude: '{{ne .Platform "ubuntu"}}'
        outcomes:
        - pass:
            when: "> 1.12.0"
            message: "dbus-user-session version > 1.12.0 is installed"
        - fail:
            message: "dbus-user-session version > 1.12.0 is required"
```

as not all installed packages comply with semantic version the analyzer
attempts to use the `when` clause as a regex if the package is installed
but does not comply with semantic versioning.