feat(collectors): Add mTLS parameters to the redis collector
For a redis collector spec targeting a redis server configured to accept
mTLS connections we need to pass in the necessary TLS parameters in order
to successfully connect to the server. Both preflight and support bundle
specs use this collector.

This change allows us to pass in the necessary TLS parameters via inlined
TLS configuration or via a secret reference.

Fixes #746