feat: allow to specify dial interface
by default the tcp port status collector listen on all interfaces but
selects the first interface found to issue the request. this commit
makes the collector to accept an interface to which we want to call.

for example:

```yaml
  collectors:
  - tcpPortStatus:
      collectorName: test
      port: 10248
      dialInterface: lo
```

this definition will make the collector listen on all interface but call
127.0.0.1:10248 to test. other example could be:

```yaml
  collectors:
  - tcpPortStatus:
      collectorName: test
      port: 10248
      dialInterface: eth0
```

this again will make the collector listen in all interfaces but it will
use the ip found on eth0 to issue the call.